move signal realated things to signal.py
